added divergence from randomness ranking (I(ne)B2) according to Amati & van Rijksbergen (2002) TOIS 20(4):357-389
git-svn-id: https://groke.otago.ac.nz/ant/ant@389 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/search_engine_accumulator.h
@@ -17,7 +17,7 @@ public:
 		using floating point arrithmatic.  The effect on precision is in the 5th decimal place (negligable) and so can be ignored.
 		For higher accuracy at an execution cost set ANT_accumulator_t to a double (or float).
 	*/
-	typedef short ANT_accumulator_t;			// short or double
+	typedef double ANT_accumulator_t;			// short or double
 
 private:
 	ANT_accumulator_t rsv;
